course_6/module_3_lab_1/requirements.txt
openai==0.28.1
uvicorn==0.27.1
fastapi==0.109.2
python-dotenv